drivers/sms_ser.c: keep the decimal digit of measured values [#1987]
The UPS reports tenths (e.g. 0x04ff for 127.9 V), but the long was divided
by 10 before becoming a double, so NUT always showed 127.00, 59.00 etc.

Developed with Claude Code (model Claude Fable 5.1).

+++ drivers/sms_ser.c
@@ -103,43 +103,43 @@ void sms_parse_results(uint8_t *rawvalues, SmsData *results) {
     memset(buf, 0, BUFFER_SIZE);
     snprintf(buf, sizeof(buf), "0x%02x%02x", (unsigned char)rawvalues[1], (unsigned char)rawvalues[2]);
     v = strtol(buf, NULL, 16); /* 16 == hex */
-    h = v / 10;
+    h = (double)v / 10;
     results->lastinputVac = h;
 
     memset(buf, 0, BUFFER_SIZE);
     snprintf(buf, sizeof(buf), "0x%02x%02x", (unsigned char)rawvalues[3], (unsigned char)rawvalues[4]);
     v = strtol(buf, NULL, 16); /* 16 == hex */
-    h = v / 10;
+    h = (double)v / 10;
     results->inputVac = h;
 
     memset(buf, 0, BUFFER_SIZE);
     snprintf(buf, sizeof(buf), "0x%02x%02x", (unsigned char)rawvalues[5], (unsigned char)rawvalues[6]);
     v = strtol(buf, NULL, 16); /* 16 == hex */
-    h = v / 10;
+    h = (double)v / 10;
     results->outputVac = h;
 
     memset(buf, 0, BUFFER_SIZE);
     snprintf(buf, sizeof(buf), "0x%02x%02x", (unsigned char)rawvalues[7], (unsigned char)rawvalues[8]);
     v = strtol(buf, NULL, 16); /* 16 == hex */
-    h = v / 10;
+    h = (double)v / 10;
     results->outputpower = h;
 
     memset(buf, 0, BUFFER_SIZE);
     snprintf(buf, sizeof(buf), "0x%02x%02x", (unsigned char)rawvalues[9], (unsigned char)rawvalues[10]);
     v = strtol(buf, NULL, 16); /* 16 == hex */
-    h = v / 10;
+    h = (double)v / 10;
     results->outputHz = h;
 
     memset(buf, 0, BUFFER_SIZE);
     snprintf(buf, sizeof(buf), "0x%02x%02x", (unsigned char)rawvalues[11], (unsigned char)rawvalues[12]);
     v = strtol(buf, NULL, 16); /* 16 == hex */
-    h = v / 10;
+    h = (double)v / 10;
     results->batterylevel = h;
 
     memset(buf, 0, BUFFER_SIZE);
     snprintf(buf, sizeof(buf), "0x%02x%02x", (unsigned char)rawvalues[13], (unsigned char)rawvalues[14]);
     v = strtol(buf, NULL, 16); /* 16 == hex */
-    h = v / 10;
+    h = (double)v / 10;
     results->temperatureC = h;
 
     byte = rawvalues[15];
